+++ typed-float-ctors/test/test.js
@@ -24,6 +24,7 @@ var tape = require( 'tape' );
 var dtypes = require( './../../typed-float-dtypes' );
 var Float64Array = require( './../../float64' );
 var Float32Array = require( './../../float32' );
+var Float16Array = require( './../../float16' );
 var Complex128Array = require( './../../complex128' );
 var Complex64Array = require( './../../complex64' );
 var isFunction = require( '@stdlib/assert/is-function' );
@@ -47,12 +48,14 @@ tape( 'the function returns typed array constructors', function test( t ) {
 	dtypes = [
 		'float64',
 		'float32',
+		'float16',
 		'complex128',
 		'complex64'
 	];
 	expected = [
 		Float64Array,
 		Float32Array,
+		Float16Array,
 		Complex128Array,
 		Complex64Array
 	];